compressed all header components into the Header component to reduce DOM space
